+++ packages/client/src/clients/preferences/interfaces.ts
@@ -8,9 +8,14 @@ export type ChannelTypePreferences = {
   [_K in ChannelType]?: boolean | ConditionalPreferenceSettings;
 };
 
+export type ChannelPreferences = {
+  [channel_id: string]: boolean | ConditionalPreferenceSettings;
+};
+
 export type WorkflowPreferenceSetting =
   | boolean
   | { channel_types: ChannelTypePreferences }
+  | { channels: ChannelPreferences }
   | ConditionalPreferenceSettings;
 
 export type WorkflowPreferences = Partial<
@@ -21,6 +26,7 @@ export interface SetPreferencesProperties {
   workflows: WorkflowPreferences;
   categories: WorkflowPreferences;
   channel_types: ChannelTypePreferences;
+  channels: ChannelPreferences;
 }
 
 export interface PreferenceSet {
@@ -28,6 +34,7 @@ export interface PreferenceSet {
   categories: WorkflowPreferences | null;
   workflows: WorkflowPreferences | null;
   channel_types: ChannelTypePreferences | null;
+  channels: ChannelPreferences | null;
 }
 
 export interface PreferenceOptions {